put expert codes from assignment 3 as base in assignment 4
