--- 4 - WS Part 3/src/Expert/ButtonValue.java
@@ -1,3 +1,0 @@
-public enum ButtonValue{
-	CLOSE
-}
