+++ 4 - WS Part 3/src/WindowSystem.java
@@ -82,7 +82,17 @@ public class WindowSystem extends GraphicsEventSystem{
             int rightBottomX = t.getRightBottomX();
             int rightBottomY = t.getRightBottomY();
 
-            //drawing window squares
+            //draw shadow first
+            if(windowManager != null){
+                super.setColor(windowManager.getShadowColor());
+                super.fillRect(t.getLeftTopX()+15 , t.getLeftTopY()+15, t.getRightBottomX()+10, t.getRightBottomY()+10);
+                super.setColor(t.getFrameColorWindow());
+                super.drawRect(leftTopX-2 , leftTopY-2, rightBottomX+1, rightBottomY+1);
+                super.setColor(Color.WHITE);
+                super.drawRect(leftTopX-1 , leftTopY-1, rightBottomX, rightBottomY);
+            }
+
+            //drawing window squares and border
             super.setColor(t.getFillColorWindow());
             super.fillRect(leftTopX , leftTopY, rightBottomX, rightBottomY);
 
@@ -96,9 +106,11 @@ public class WindowSystem extends GraphicsEventSystem{
                     rectangleComponent.getY() + rectangleComponent.getHeight());
 
                 //draw the title of the window
-                rectangleComponent.setString(t.getTitle());
-                super.setColor(Color.black);
-                super.drawString(rectangleComponent.getString(), t.getLeftTopX() + 10, t.getLeftTopY() + 15);
+                rectangleComponent.setString(rectangleComponent.getString());
+                super.setColor(rectangleComponent.getStringColor());
+                super.drawString(rectangleComponent.getString(), 
+                    rectangleComponent.getX() + rectangleComponent.getStringPaddingLeft(), 
+                    rectangleComponent.getY() + rectangleComponent.getStringPaddingTop());
             }
         }
     }
